Put everything into sections and clean up semantics
