+++ pages/index.tsx
@@ -2,19 +2,17 @@
 // @ts-nocheck
 
 import CallToAction from '@/components/call-to-action';
-import Heading from '@/components/fixed-krado-components/Heading';
 import Layout from '@/components/layout';
 import FullScreenVideo from '@/components/work/full-screen-video';
-import ProjectCard from '@/components/work/project-card';
 import ReelVideo from '@/components/work/reel-video';
-import { Box, Container, Flex, Grid } from 'krado-react';
+import { Box, Flex } from 'krado-react';
 import { useState } from 'react';
 import Shade, { ShadeButton } from '@/components/shade';
 import { motion, useScroll, useSpring, useTransform } from 'framer-motion';
-import { useRef } from 'react';
-import Text from '@/components/fixed-krado-components/Text';
 import client from '@/tina/__generated__/client';
 import { useTina } from 'tinacms/dist/react';
+import ProjectsSection from '@/components/work/projects-section';
+import ServicesSection from '@/components/work/services-section';
 
 export default function Work(props) {
   const { data } = useTina({
@@ -28,10 +26,7 @@ export default function Work(props) {
   const [isVideoHidden, setIsVideoHidden] = useState(true);
   const [isCovered, setIsCovered] = useState(false);
 
-  const ref = useRef(null);
-
   const { scrollYProgress } = useScroll({
-    target: ref,
     offset: ['0 1', '0.45 1']
   });
 
@@ -43,8 +38,6 @@ export default function Work(props) {
   const workScrollSpring = useSpring(workScroll, { damping: 20 });
   const videoScroll = useTransform(scrollYProgress, [0, 1], [1, 0.9]);
   const borderScroll = useTransform(scrollYProgress, [0, 1], [0, 80]);
-  const workHeaderScroll = useTransform(scrollYProgress, [0.8, 1], [0, 1]);
-  const workHeaderScrollSpring = useSpring(workHeaderScroll, { damping: 20 });
 
   return (
     <Layout disableScroll={!isVideoHidden}>
@@ -112,99 +105,20 @@ export default function Work(props) {
             backgroundColor: 'background',
             flexDirection: 'column',
             borderRadius: ['40px', '80px'],
-
             zIndex: 2
           }}
           style={{ scale: workScrollSpring }}
         >
-          <Container>
-            <div ref={ref}>
-              <Heading
-                as={motion.h2}
-                variant='display.display'
-                id='work'
-                sx={{
-                  marginTop: 5,
-                  marginBottom: 3,
-                  textAlign: 'center',
-                  scrollMarginTop: '3.8em'
-                }}
-                style={{
-                  scale: workScrollSpring,
-                  opacity: workHeaderScrollSpring
-                }}
-              >
-                Featured work
-              </Heading>
-
-              <Text
-                as={motion.h2}
-                variant='body.summary'
-                sx={{
-                  marginBottom: 5,
-                  textAlign: 'center',
-                  fontWeight: 400
-                }}
-                style={{
-                  scale: workScrollSpring,
-                  opacity: workHeaderScrollSpring
-                }}
-              >
-                Ut nunc, dui sit sit nisl, cras velit lorem. Laoreet gravida
-                adipiscing augue sit.
-              </Text>
-
-              <Grid
-                sx={{
-                  gap: 4,
-                  gridTemplateColumns: ['1fr', null, '1fr 1fr']
-                }}
-              >
-                {projectsList.map(({ project }) => (
-                  <ProjectCard
-                    title={project.title}
-                    href={`/projects/${project._sys.filename}`}
-                    src={project.coverImage}
-                    key={project._sys.filename}
-                  />
-                ))}
-              </Grid>
-            </div>
-          </Container>
-
-          {/* <Container>
-            <Heading
-              variant="body.pretext"
-              sx={{
-                marginBottom: [2, 3]
-              }}
-            >
-              Areas of expertise
-            </Heading>
-            <HorizontalCard
-              imageAlt="hi"
-              imageSrc="/turntable-thumbnail.png"
-              label="Motion design"
-            />
-            <HorizontalCard
-              imageAlt="hi"
-              imageSrc="/turntable-thumbnail.png"
-              label="Lorem ipsum"
-            />
-            <HorizontalCard
-              imageAlt="hi"
-              imageSrc="/turntable-thumbnail.png"
-              label="Lorem ipsum"
-            />
-          </Container> */}
+          <ProjectsSection projects={projectsList} />
+          <ServicesSection />
         </Flex>
-        <motion.section style={{ opacity: scrollCtaProgress }}>
+
+        <motion.footer style={{ opacity: scrollCtaProgress }}>
           <CallToAction
             title="Let's connect"
             text='Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris id auctor neque, eu dictum urna.'
           />
-        </motion.section>
-        <Box />
+        </motion.footer>
       </Flex>
     </Layout>
   );
@@ -220,7 +134,6 @@ export const getStaticProps = async () => {
       data,
       query,
       variables
-      //myOtherProp: 'some-other-data',
     }
   };
 };
